update deploy js

+++ controllers/upgrade.go
@@ -14,12 +14,19 @@ const upgradePwd = "kid1234"
 
 func upgrade(c *gin.Context) {
 	if c.Query("pwd") == upgradePwd {
-		ExecCommand("git pull")
-		api.Success(c, "upgrade成功 至"+ExecCommand("git rev-parse HEAD")) // 没有数据
-		go func() {
-			time.Sleep(1)
-			ExecCommand("ps -ef|grep go-build|grep -v grep|cut -c 9-15|xargs kill -9")
-		}()
+		if c.Query("type") == "api" || c.Query("type") == "" {
+			ExecCommand("git pull")
+			api.Success(c, "upgrade成功 至"+ExecCommand("git rev-parse HEAD")) // 没有数据
+			go func() {
+				time.Sleep(1)
+				ExecCommand("ps -ef|grep go-build|grep -v grep|cut -c 9-15|xargs kill -9")
+			}()
+		}
+		if c.Query("type") == "js" {
+			ExecCommand("sshpass -p Zoomus123 ssh  -o StrictHostKeyChecking=no  root@49.234.73.182 '/opt/workdoc/buildjs.sh'")
+			api.Success(c, "upgrade成功 至"+ExecCommand("cd /opt/nginxdocker/mulkids-cms-pro;git rev-parse HEAD")) // 没有数据
+		}
+
 	} else {
 		api.Fail(c, http.StatusBadRequest, "密码错误")
 	}
